refactor: Add dummy rows at start/end of game
These dummy rows assign both teams a win probability of 0.5 at the
beginning and end of the game to make the plot shading cleaner.
There is also a dummy row to assign a wp of 1 to the winning team at
the end of the game since the last play is usually not at time 0 and
has a wp of 0.99.