refactor: Make win prob plot time scale dynamic
Uses a reference line at the end of the game to allow Tableau to
automatically set the range for the win probability plot but ensures
that both overlaid plots have the same scale.

Automatic scaling is needed to account for games that go to overtime
and don't have a fixed end of game.